fix duplicated city reference
