--- ProjectCaravanGame/Assets/Modules/Controllers/City.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8a29c69a74c629d4db3d05b2b6506157
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
