--- ProjectCaravanGame/Assets/Modules/Testing/RoadMovementTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8ee779d90aee5f54aa72023ed74ac8e5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
